pyproject.toml
[tool.poetry]
name = "functional-gpt"
version = "0.1.0"
description = ""
authors = ["Phate <phate334@gmail.com>"]
readme = "README.md"
packages = [{include = "fgpt"}]

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.7.0"}
pydantic = {extras = ["dotenv"], version = "^1.10.6"}
openai = "^0.27.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fgpt = "fgpt.cli:app"